Avancement affichage

+++ src/vue/Point.java
@@ -1,27 +1,27 @@
 package vue;
 
 public class Point {
-    protected int x;
-    protected int y;
+    protected float x;
+    protected float y;
 
-    public Point(int x, int y) {
+    public Point(float x, float y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(float y) {
         this.y = y;
     }
 }
